chore(storage): simplify storage ports and web adapter wiring
