+++ src/app/ports/key-value-storage.port.ts
@@ -1,0 +1,8 @@
+export interface KeyValueStoragePort {
+  get(key: string): string | null;
+  set(key: string, value: string): void;
+  remove(key: string): void;
+  keys(): string[];
+  length(): number;
+  clear(): void;
+}
