+++ src/app/ports/storage-provider.port.ts
@@ -1,0 +1,6 @@
+import { StoreType } from '@src/domain/entities/store-type';
+import { KeyValueStoragePort } from '@src/app/ports/key-value-storage.port';
+
+export interface StorageProviderPort {
+  get(type: StoreType): KeyValueStoragePort;
+}
