+++ src/app/ports/storage.port.ts
@@ -1,12 +1,6 @@
-import { StoreType } from '@src/domain/entities/store-type';
+import { KeyValueStoragePort } from '@src/app/ports/key-value-storage.port';
 
-export interface KeyValueStoragePort {
-  get(key: string): string | null;
-  set(key: string, value: string): void;
-  remove(key: string): void;
-  keys(): string[];
-}
-
-export interface StorageProviderPort {
-  get(type: StoreType): KeyValueStoragePort;
+export interface StoragePort {
+  local: KeyValueStoragePort;
+  session: KeyValueStoragePort;
 }
