+++ src/interface/index.ts
@@ -1,12 +1,9 @@
 import { StrictStoreService } from '@src/app/strict-store.service';
 import { createKey } from '@src/domain/policies/create-key';
+import { WebStorageProvider } from '@src/infrastructure/providers/web-storage.provider';
 import { strictJson } from '@src/infrastructure/serialization/strict-store.adapter';
-import { BrowserStorageProvider } from '@src/infrastructure/web/web-storage.adapter';
 
-const createWebStrictStore = () =>
-  new StrictStoreService(new BrowserStorageProvider(), strictJson);
-
-const StrictStore = createWebStrictStore();
+const StrictStore = new StrictStoreService(new WebStorageProvider(), strictJson);
 
 export {
   StrictStore,
